# openapi3.yaml
openapi: 3.0.1
info:
  title: 3d-ops-trigger
  description: Entry point for 3D operations — validates requests and triggers Jobnik jobs
  version: 1.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
paths:
  /jobOperations/ingestion:
    post:
      operationId: createIngestion
      tags:
        - jobOperations
      summary: Create a new ingestion job to invoke a new model ingestion flow
      description: Validates the request (light & fast) and creates a Jobnik ingestion job.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ingestionPayload'
      responses:
        201:
          description: Ingestion job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /jobOperations/delete:
    post:
      operationId: createDelete
      tags:
        - jobOperations
      summary: Create a new delete job to invoke a delete flow
      description: Validates the record can be deleted and creates a Jobnik delete job.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/deletePayload'
      responses:
        200:
          description: Delete job created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /jobStatus/{jobId}:
    get:
      operationId: getJobStatus
      tags:
        - jobStatus
      summary: Get the status and progress of a job
      parameters:
        - $ref: '#/components/parameters/jobId'
      responses:
        200:
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/jobStatusResponse'
        404:
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /models/validate:
    post:
      operationId: validateModel
      tags:
        - models
      summary: Validate an ingestion request without creating a job
      description: Runs the full ingestion validation and returns the result without triggering a Jobnik job.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ingestionPayload'
      responses:
        200:
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validationResultResponse'
  /models/canDelete/{recordId}:
    get:
      operationId: canDeleteModel
      tags:
        - models
      summary: Validate whether a record can be deleted
      description: Runs the delete validation and returns the result without creating a job.
      parameters:
        - $ref: '#/components/parameters/recordId'
      responses:
        200:
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/validationResultResponse'
  /metadata/{identifier}:
    patch:
      operationId: updateMetadata
      tags:
        - metadata
      summary: Update metadata of a model
      parameters:
        - $ref: '#/components/parameters/identifier'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updatePayload'
      responses:
        200:
          description: Metadata updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ackResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
  /metadata/status/{identifier}:
    patch:
      operationId: updateMetadataStatus
      tags:
        - metadata
      summary: Update the publish/unpublish status of a model
      parameters:
        - $ref: '#/components/parameters/identifier'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/statusPayload'
      responses:
        200:
          description: Status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ackResponse'
        400:
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
        404:
          description: Record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
security:
  - {}
components:
  parameters:
    identifier:
      name: identifier
      in: path
      required: true
      description: The record identifier
      schema:
        type: string
    recordId:
      name: recordId
      in: path
      required: true
      description: The record identifier
      schema:
        type: string
    jobId:
      name: jobId
      in: path
      required: true
      description: The job identifier
      schema:
        type: string
  schemas:
    error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    ingestionPayload:
      type: object
      required:
        - modelPath
        - tilesetFilename
        - metadata
      properties:
        modelPath:
          type: string
          description: Path on the shared storage to the 3DTiles folder or 3TZ archive
        tilesetFilename:
          type: string
          description: The tileset entry name (e.g. tileset.json)
        metadata:
          type: object
          description: 3D record metadata (business validation applied downstream)
          additionalProperties: true
    deletePayload:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          description: The record identifier to delete
    updatePayload:
      type: object
      description: Partial metadata fields to update
      additionalProperties: true
    statusPayload:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - PUBLISHED
            - UNPUBLISHED
    jobResponse:
      type: object
      required:
        - jobId
        - status
      properties:
        jobId:
          type: string
        status:
          type: string
    ackResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
    jobStatusResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
        percentage:
          type: number
    validationResultResponse:
      type: object
      required:
        - isValid
      properties:
        isValid:
          type: boolean
        message:
          type: string